feat: アレルゲンおよび宗教的制約に関する定数とマップを再エクスポートし、関連するIDマップを追加
