+++ functions/src/constants.ts
@@ -1,0 +1,7 @@
+// プロジェクトルートの定数やマップを再エクスポート
+
+export {
+  religiousRestrictionNameList,
+  religiousRestrictionNameToIdMap,
+} from "../../types/religious"
+export { allergenNameList, allergyNameToIdMap } from "../../types/allergy"
